finish status
